agentic/src/app/agent/state.py
from typing import Annotated

from langgraph.graph.message import add_messages
from typing_extensions import TypedDict


class AgentState(TypedDict):
    messages: Annotated[list, add_messages]
    user_id: str
    session_id: str
    reflection_count: int
    blocked: bool
    block_reason: str | None
